+++ src/main/java/com/example/boardproject/entity/Board.java
@@ -4,6 +4,7 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import com.example.boardproject.domain.UploadFile;
+
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -15,20 +16,20 @@ import java.util.List;
 @ToString
 @Builder
 @EntityListeners(value = {AuditingEntityListener.class})
-public class Board{
+public class Board {
 
     @Id
-    @Column(name="bId")
+    @Column(name = "bId")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long bId;
+    private int bId;
 
-    @Column(length = 30, name="bTitle",nullable = false)
+    @Column(length = 30, name = "bTitle", nullable = false)
     private String bTitle;
 
-    @Column(length = 20, name="bWriter",nullable = false)
+    @Column(length = 20, name = "bWriter", nullable = false)
     private String bWriter;
 
-    @Column(length = 100, name="bContent")
+    @Column(length = 100, name = "bContent")
     private String bContent;
 
     @ElementCollection
@@ -36,18 +37,18 @@ public class Board{
             joinColumns = @JoinColumn(name = "fId"))
     private List<UploadFile> pImageFiles;
 
-    @Column(length = 20,name="bPw", nullable = false)
+    @Column(length = 20, name = "bPw", nullable = false)
     private String bPw;
 
     @CreationTimestamp
-    @Column(updatable = false, name="bDate")
+    @Column(updatable = false, name = "bDate")
     private LocalDateTime bDate;
 
     @ManyToOne
     @JoinColumn(name = "pId")
     private Product pId;
 
-    public Board(String bTitle, String bWriter, String bContent, List<UploadFile> pImageFiles, String bPw,Product pid) {
+    public Board(String bTitle, String bWriter, String bContent, List<UploadFile> pImageFiles, String bPw, Product pid) {
         this.bTitle = bTitle;
         this.bWriter = bWriter;
         this.bContent = bContent;
