+++ src/main/java/com/example/boardproject/entity/Product.java
@@ -14,7 +14,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long pId;
+    private int pId;
 
     @Column(length = 10)
     private String pName;
