+++ src/main/java/com/example/boardproject/service/BoardServiceImpl.java
@@ -1,7 +1,6 @@
 package com.example.boardproject.service;
 
 import com.example.boardproject.dto.BoardDTO;
-import com.example.boardproject.dto.BoardWriteDTO;
 import com.example.boardproject.dto.PageRequestDTO;
 import com.example.boardproject.dto.PageResultDTO;
 import com.example.boardproject.entity.Board;
@@ -29,13 +28,13 @@ public class BoardServiceImpl implements BoardService {
         this.searchRepository = searchRepository;
     }
 
-    public PageResultDTO<BoardDTO, Board> getList(Long pId, PageRequestDTO pageRequestDTO) {
+    public PageResultDTO<BoardDTO, Board> getList(int pId, PageRequestDTO pageRequestDTO) {
         PageRequest pageRequest = pageRequestDTO.getPageable(Sort.by("bId").descending());
 
-        Page<Board> result = searchRepository.findByPId(pId,pageRequest);
+        Page<Board> result = searchRepository.findByPId(pId, pageRequest);
 
         Function<Board, BoardDTO> fn = (entity -> entityToDto(entity));
 
-        return new PageResultDTO<>(result,fn);
+        return new PageResultDTO<>(result, fn);
     }
 }
